add logic for create kiam action
