+++ cmd/action/create/kiam/awsapicall/job.go
@@ -1,4 +1,4 @@
-package ac009
+package awsapicall
 
 import (
 	"fmt"
