--- cmd/cl001/ac009/zz_generated.error.go
@@ -1,21 +1,0 @@
-package ac009
-
-import "github.com/giantswarm/microerror"
-
-var invalidConfigError = &microerror.Error{
-	Kind: "invalidConfigError",
-}
-
-// IsInvalidConfig asserts invalidConfigError.
-func IsInvalidConfig(err error) bool {
-	return microerror.Cause(err) == invalidConfigError
-}
-
-var invalidFlagsError = &microerror.Error{
-	Kind: "invalidFlagsError",
-}
-
-// IsInvalidFlags asserts invalidFlagsError.
-func IsInvalidFlags(err error) bool {
-	return microerror.Cause(err) == invalidFlagsError
-}
