--- cmd/cl001/ac009/zz_generated.flag.go
@@ -1,13 +1,0 @@
-package ac009
-
-import "github.com/spf13/cobra"
-
-type flag struct {
-}
-
-func (f *flag) Init(cmd *cobra.Command) {
-}
-
-func (f *flag) Validate() error {
-	return nil
-}
