+++ cmd/cl004/zz_generated.command.go
@@ -10,7 +10,6 @@ import (
 
 	"github.com/giantswarm/awscnfm/v12/cmd/cl004/ac000"
 	"github.com/giantswarm/awscnfm/v12/cmd/cl004/ac008"
-	"github.com/giantswarm/awscnfm/v12/cmd/cl004/ac010"
 	"github.com/giantswarm/awscnfm/v12/cmd/cl004/ac011"
 	"github.com/giantswarm/awscnfm/v12/cmd/cl004/ac012"
 	"github.com/giantswarm/awscnfm/v12/cmd/cl004/ac014"
@@ -63,20 +62,6 @@ func New(config Config) (*cobra.Command, error) {
 		}
 
 		ac008Cmd, err = ac008.New(c)
-		if err != nil {
-			return nil, microerror.Mask(err)
-		}
-	}
-
-	var ac010Cmd *cobra.Command
-	{
-		c := ac010.Config{
-			Logger: config.Logger,
-			Stderr: config.Stderr,
-			Stdout: config.Stdout,
-		}
-
-		ac010Cmd, err = ac010.New(c)
 		if err != nil {
 			return nil, microerror.Mask(err)
 		}
@@ -144,7 +129,6 @@ func New(config Config) (*cobra.Command, error) {
 
 	c.AddCommand(ac000Cmd)
 	c.AddCommand(ac008Cmd)
-	c.AddCommand(ac010Cmd)
 	c.AddCommand(ac011Cmd)
 	c.AddCommand(ac012Cmd)
 	c.AddCommand(ac014Cmd)
