--- pkg/action/cl001/ac009/error.go
@@ -1,34 +1,0 @@
-package ac009
-
-import "github.com/giantswarm/microerror"
-
-// executionFailedError is an error type for situations where Resource execution
-// cannot continue and must always fall back to operatorkit.
-//
-// This error should never be matched against and therefore there is no matcher
-// implement. For further information see:
-//
-//     https://github.com/giantswarm/fmt/blob/master/go/errors.md#matching-errors
-//
-var executionFailedError = &microerror.Error{
-	Kind: "executionFailedError",
-}
-
-var notFoundError = &microerror.Error{
-	Kind: "notFoundError",
-}
-
-// IsNotFound asserts notFoundError.
-func IsNotFound(err error) bool {
-	return microerror.Cause(err) == notFoundError
-}
-
-var tooManyCRsError = &microerror.Error{
-	Kind: "tooManyCRsError",
-	Desc: "There is only a single AWSCluster CR allowed with the current implementation.",
-}
-
-// IsTooManyCRsError asserts tooManyCRsError.
-func IsTooManyCRsError(err error) bool {
-	return microerror.Cause(err) == tooManyCRsError
-}
